``` feat(vector-store): 将向量存储从Chroma切换到FAISS以支持Windows环境
为解决Chroma在Windows上需要MSVC编译的问题，改为使用FAISS作为向量存储。
FAISS提供预编译包，体积小、速度快，更适合Windows部署。

BREAKING CHANGE: 向量存储后端从Chroma更换为FAISS，需要重新初始化数据。
```

+++ app/rag/chain.py
@@ -100,7 +100,7 @@ async def chat(
             "doc_id": doc.metadata.get("doc_id", ""),
             "filename": doc.metadata.get("filename", "unknown"),
             "content": doc.page_content[:500],  # 截断避免返回太大
-            "score": round(score, 4),
+            "score": float(round(float(score), 4)),  # numpy float32 -> python float
         }
         for doc, score in results
     ]
